+++ include/accfft_gpuf.h
@@ -35,7 +35,7 @@
 
 struct accfft_plan_gpuf {
 	int N[3];
-	int alloc_max;
+	size_t alloc_max;
 	Mem_Mgr_gpu<float> * Mem_mgr;
 	T_Plan_gpu<float> * T_plan_1;
 	T_Plan_gpu<float> * T_plan_2;
@@ -94,7 +94,7 @@ struct accfft_plan_gpuf {
 
 };
 
-int dfft_get_local_size_gpuf(int N0, int N1, int N2, int * isize, int * istart,
+size_t dfft_get_local_size_gpuf(int N0, int N1, int N2, int * isize, int * istart,
 		MPI_Comm c_comm);
 
 void accfft_cleanup_gpuf();
@@ -117,11 +117,11 @@ void accfft_execute_c2c_gpuf(accfft_plan_gpuf* plan, int direction,
 accfft_plan_gpuf* accfft_plan_dft_3d_c2c_gpuf(int * n, Complexf * data_d,
 		Complexf * data_out_d, MPI_Comm c_comm,
 		unsigned flags = ACCFFT_MEASURE);
-int accfft_local_size_dft_c2c_gpuf(int * n, int * isize, int * istart,
+size_t accfft_local_size_dft_c2c_gpuf(int * n, int * isize, int * istart,
 		int * osize, int *ostart, MPI_Comm c_comm);
 accfft_plan_gpuf* accfft_plan_dft_3d_r2c_gpuf(int * n, float * data_d,
 		float * data_out_d, MPI_Comm c_comm, unsigned flags = ACCFFT_MEASURE);
-int accfft_local_size_dft_r2c_gpuf(int * n, int * isize, int * istart,
+size_t accfft_local_size_dft_r2c_gpuf(int * n, int * isize, int * istart,
 		int * osize, int *ostart, MPI_Comm c_comm);
 
 template<typename T, typename Tc>
@@ -132,7 +132,7 @@ void accfft_execute_c2r_gpu_t(accfft_plan_gpuf* plan, Tc* data, T* data_out,
 		double * timer = NULL, std::bitset<3> XYZ = 111);
 
 template<typename T>
-int accfft_local_size_dft_r2c_gpu_t(int * n, int * isize, int * istart,
+size_t accfft_local_size_dft_r2c_gpu_t(int * n, int * isize, int * istart,
 		int * osize, int *ostart, MPI_Comm c_comm);
 
 // templates for execution in z direction only
